+++ types/obj_darwin.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"debug/macho"
+	"fmt"
 	"os"
 )
 
@@ -15,4 +16,19 @@ func init() {
 	if err := initMachOFunctionTables(f); err != nil {
 		panic(err)
 	}
+}
+
+func initMachOFunctionTables(f *macho.File) error {
+	pclntab := f.Section("__gopclntab")
+	pclntabData, err := readAll(pclntab, pclntab.Size)
+	if err != nil {
+		return fmt.Errorf("cannot read pclntab: %w", err)
+	}
+	symtab := f.Section("__gosymtab")
+	symtabData, err := readAll(symtab, symtab.Size)
+	if err != nil {
+		return fmt.Errorf("cannot read symtab: %w", err)
+	}
+	initFunctionTables(pclntabData, symtabData)
+	return nil
 }
